pyproject.toml
[project]
name = "graph-tools-for-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kaleido==0.1.0.post1 ; sys_platform == 'win32'",
    "kaleido==0.2.1 ; sys_platform == 'linux'",
    "openai-agents>=0.6.4",
    "plotly>=6.5.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pywin32>=311",
    "viktor>=14.26.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
